Renamed exceptionmapper and models packages
