+++ core/src/main/java/com/cerner/beadledom/core/BeadledomModule.java
@@ -6,9 +6,9 @@ import com.cerner.beadledom.configuration.BeadledomConfigurationModule;
 import com.cerner.beadledom.health.HealthModule;
 import com.cerner.beadledom.jackson.JacksonModule;
 import com.cerner.beadledom.jaxrs.JaxRsModule;
-import com.cerner.beadledom.jaxrs.exceptionmapper.FailureExceptionMapper;
-import com.cerner.beadledom.jaxrs.exceptionmapper.ThrowableExceptionMapper;
-import com.cerner.beadledom.jaxrs.exceptionmapper.WebApplicationExceptionMapper;
+import com.cerner.beadledom.jaxrs.exceptionmapping.FailureExceptionMapper;
+import com.cerner.beadledom.jaxrs.exceptionmapping.ThrowableExceptionMapper;
+import com.cerner.beadledom.jaxrs.exceptionmapping.WebApplicationExceptionMapper;
 import com.cerner.beadledom.jaxrs.provider.FilteringJacksonJsonProvider;
 import com.cerner.beadledom.stagemonitor.StagemonitorModule;
 import com.cerner.beadledom.swagger.SwaggerModule;
