+++ json-common/pom.xml
@@ -34,5 +34,4 @@
             <artifactId>swagger-annotations</artifactId>
         </dependency>
     </dependencies>
-
 </project>
